% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_train.R
\docType{data}
\name{data_train}
\alias{data_train}
\title{7 sets of PRSs for training data set and target phenotype}
\format{
A data frame for training dataset:
\describe{
\item{V1}{PRS1, for bin1}
\item{V2}{PRS2, for bin1}
\item{V3}{PRS3, for bin1}
\item{V4}{PRS4, for bin1}
\item{V5}{PRS5, for bin1}
\item{V6}{PRS6, for bin1}
\item{V7}{PRS7, for bin1}
\item{phenotype}{Target Phenotype, value}

}
}
\usage{
data_train
}
\description{
A dataset containing 7 sets of PRSs for training data set and target phenotype
}
\keyword{datasets}
