% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_valid.R
\docType{data}
\name{data_valid}
\alias{data_valid}
\title{7 sets of PRSs for validation dataset and target phenotype}
\format{
A data frame for validation dataset:
\describe{
\item{V1}{Feature 1 (or PRS1 constructed using the first subset of SNPs from GWAS summary statistics)}
\item{V2}{Feature 2 (or PRS2 constructed using the second subset of SNPs from GWAS summary statistics)}
\item{V3}{Feature 3 (or PRS3 constructed using the third subset of SNPs from GWAS summary statistics)}
\item{V4}{Feature 4 (or PRS4 constructed using the fourth subset of SNPs from GWAS summary statistics)}
\item{V5}{Feature 5 (or PRS5 constructed using the fifth subset of SNPs from GWAS summary statistics)}
\item{V6}{Feature 6 (or PRS6 constructed using the sixth subset of SNPs from GWAS summary statistics)}
\item{V7}{Feature 7 (or PRS7 constructed using the seventh subset of SNPs from GWAS summary statistics)}
\item{phenotype}{Phenotypic values}

}
}
\usage{
data_valid
}
\description{
A dataset containing 7 sets of PRSs for validation dataset and target phenotype
}
\keyword{datasets}
