% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{fourth_root_transform}
\alias{fourth_root_transform}
\title{Fourth root transform data}
\usage{
fourth_root_transform(ms)
}
\arguments{
\item{ms}{}
}
\value{
ms
}
\description{
Fourth root transform data
}
\examples{
fourth_root_transform(ms)
}
