% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_preprocessed.R
\name{load_data}
\alias{load_data}
\title{Load data returned by XCMS getPeakTable function}
\usage{
load_data(path)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A list of two dataframes
}
\description{
This function loads a csv as returned by getPeakTable (xcms). It wrangles
the data and returns two dataframes: a peak meta data data frame
describing mz, rt, isotopes, etc. and a peaks table with intensities and
observation ids.
The output is used in other ezmz functions.
}
