% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_part1.r
\name{model_combination}
\alias{model_combination}
\title{model_combination function
This function will generate PRS based on all possible combianations of model.
The total number of models required to explore the combinations of these 'n'
features can be calculated by summing the combinations for each possible
number of features, ranging from 1 to 'n' (?\link{C(n,i)}).
where C(n,k)  represents the binomial coefficient or "n choose k,"
with n denoting the total number of features and k indicating
the number of features to include in each model.}
\usage{
model_combination(data_train, data_valid, data_test, mv)
}
\arguments{
\item{data_train}{This is the matrix for training dataset}

\item{data_valid}{This is the matrix for validation dataset}

\item{data_test}{This is the matrix for test dataset}

\item{mv}{The total number of columns in data_train/data_valid/data_test}
}
\value{
This function will generate all possible model outcomes for validation and test dataset
\item{}{}
}
\description{
model_combination function
This function will generate PRS based on all possible combianations of model.
The total number of models required to explore the combinations of these 'n'
features can be calculated by summing the combinations for each possible
number of features, ranging from 1 to 'n' (?\link{C(n,i)}).
where C(n,k)  represents the binomial coefficient or "n choose k,"
with n denoting the total number of features and k indicating
the number of features to include in each model.
}
\examples{
data_train <- data_train
data_train  <- data_train
data_test  <- data_test
mv=8
model_combination(data_train,data_valid,data_test,mv)
}
\keyword{All}
\keyword{combinations}
\keyword{model}
\keyword{possible}
