% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_part1.r
\name{model_configuration}
\alias{model_configuration}
\title{model_configuration function}
\usage{
model_configuration(data_train, data_valid, mv)
}
\arguments{
\item{data_train}{This is the matrix for training dataset}

\item{data_valid}{This is the matrix for validation dataset}

\item{mv}{The total number of columns in data_train/data_valid}
}
\value{
This function will generate all possible model outcomes for validation and test dataset
\item{}{}
}
\description{
This function will generate features (e.g. PRSs) based on all possible combinations of model.
The total number of models required to explore the combinations of these 'n'
features can be calculated by summing the combinations for each possible
number of features, ranging from 1 to 'n' (C(n,i)).
where C(n,k)  represents the binomial coefficient or "n choose k,"
with n denoting the total number of features and k indicating
the number of features to include in each model.
}
\examples{
data_train <- data_train
data_valid  <- data_valid
mv=8
out=model_configuration(data_train,data_valid,mv)
}
\keyword{All}
\keyword{combinations}
\keyword{model}
\keyword{possible}
