% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_part2.r
\name{model_evaluation}
\alias{model_evaluation}
\title{model_evaluation function}
\usage{
model_evaluation(dat, mv, tn, prev, pthreshold = 0.05, method = "R2ROC")
}
\arguments{
\item{dat}{This is the matrix for all the combinations of the model}

\item{mv}{The total number of columns in data_train/data_valid}

\item{tn}{The total no of best models to be identified}

\item{prev}{The prevalence of disease in the data}

\item{pthreshold}{The P value threshold for the significance level}

\item{method}{The methods to be used to evaluate models}
}
\value{
This function will generate all possible model outcomes for validation and test dataset
\item{}{}
}
\description{
This function will identify the best model in the validation and test dataset.
}
\examples{
#dat <- predict_validation
#mv=8
#tn=15
#prev=0.047
#model_evaluation(dat,mv,tn,prev)
}
\keyword{Identify}
\keyword{best}
\keyword{models}
