% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{normalize_batch}
\alias{normalize_batch}
\title{Z-standardization of batches}
\usage{
normalize_batch(ms, batch_column = "BATCH")
}
\arguments{
\item{ms}{}
}
\value{
ms
}
\description{
Z-standardization of batches
}
\examples{
normalize_batch(ms)
}
