% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{rm_feature_batch_inflated_zeros}
\alias{rm_feature_batch_inflated_zeros}
\title{Remove features inflated with zeros (batch-wise)}
\usage{
rm_feature_batch_inflated_zeros(ms)
}
\arguments{
\item{ms}{}
}
\value{
ms
}
\description{
This might help avoiding batch effect and identification of batches through signals with binary (on/off) behaviour
}
\examples{
rm_feature_batch_inflated_zeros(ms)
}
