% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{rm_feature_extreme_values}
\alias{rm_feature_extreme_values}
\title{Removes outliers by PCA}
\usage{
rm_feature_extreme_values(ms)
}
\arguments{
\item{ms}{}
}
\value{
ms
}
\description{
Samples are removed if: abs(x - median(x)) > (1.5 * quantile(x,0.95)-quantile(x,0.05)), where x is the PCA scores.
Scores from the first 12 components are used.
}
\details{
May be applied before row normalization but after removal of extreme features.
}
\examples{
rm_sample_pca_outliers(ms)
}
