% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{rm_sample_min_batch_occurence}
\alias{rm_sample_min_batch_occurence}
\title{Removes observations if their batches are below a given threshold}
\usage{
rm_sample_min_batch_occurence(ms, min_occurence = 5)
}
\arguments{
\item{ms}{}
}
\value{
ms
}
\description{
Is usable if preceeded by batch normalization
}
\examples{
rm_sample_min_batch_occurence(ms)
}
