% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{robust_row_normalize}
\alias{robust_row_normalize}
\title{Robust row normalization}
\usage{
robust_row_normalize(ms)
}
\arguments{
\item{ms}{}
}
\value{
ms
}
\description{
Only stable features used to normalize the rows (each sample)
Corrects matrix effects
}
\examples{
robust_row_normalize(ms)
}
