% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{select_most_variable_features}
\alias{select_most_variable_features}
\title{Select top n most variable features}
\usage{
select_most_variable_features(ms, n = 500)
}
\arguments{
\item{ms}{}
}
\value{
ms
}
\description{
Effective for model screening to avoid heavy work loads. Often top 500 contains most of the signal.
}
\examples{
select_most_variable_features(ms)
}
