% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_part1.r
\name{model_configuration}
\alias{model_configuration}
\title{model_configuration function}
\usage{
model_configuration(data_train, data_valid, mv)
}
\arguments{
\item{data_train}{This is the matrix for the training dataset}

\item{data_valid}{This is the matrix for the validation dataset}

\item{mv}{The total number of columns in data_train/data_valid}
}
\value{
This function will generate all possible model outcomes for validation and test dataset
}
\description{
This function will generate features (e.g. PRSs) based on all possible combinations of model.
The total number of models required to explore the combinations of these 'n'
features can be calculated by summing the combinations for each possible
number of features, ranging from 1 to 'n' (C(n,i)).
where C(n,k)  represents the binomial coefficient or "n choose k,"
with n denoting the total number of features and k indicating
the number of features to include in each model.
}
\examples{
\donttest{
data_train <- data_train
data_valid  <- data_valid
mv=8
out=model_configuration(data_train,data_valid,mv)
#This process will produce predicted values for the validation datasets,
#corresponding to each model configuration trained on the training dataset.
#The outcome of this function will yield variables named 'predict_validation'
#and 'total_model_configurations.
#To print the outcomes run out$predict_validation and out$total_model_configurations.
#For details (see https://github.com/mommy003/MSML). 
}
}
\keyword{All}
\keyword{combinations}
\keyword{model}
\keyword{possible}
