% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_part2.r
\name{model_evaluation}
\alias{model_evaluation}
\title{model_evaluation function}
\usage{
model_evaluation(dat, mv, tn, prev)
}
\arguments{
\item{dat}{This is the matrix for all the combinations of model}

\item{mv}{The total number of columns in data_train/data_valid/data_test}

\item{tn}{The total no of best models to be identified}

\item{prev}{The prevalance of disease in the data}
}
\value{
This function will generate all possible model outcomes for validation and test dataset
\item{}{}
}
\description{
This function will identify best model in validation and test dataset.
}
\examples{
dat <- read.table("predict_test_models")
mv=8
tn=15
prev=0.047
model_evaluation(dat,mv,tn,prev)
}
\keyword{Identify}
\keyword{best}
\keyword{models}
